refactor: split vision helpers from whatsapp media
