+++ src/core/process-group-message.ts
@@ -10,7 +10,7 @@ import { checkRateLimit, recordResponse } from '../middleware/rate-limit.js';
 import { recordBotResponse } from '../middleware/stats.js';
 import { queueRetry } from '../middleware/retry.js';
 import type { MessageContext } from '../ai/persona.js';
-import type { VisionImage } from '../features/media.js';
+import type { VisionImage } from './vision.js';
 import type { PlatformMessenger } from './platform-messenger.js';
 
 export interface ProcessGroupMessageParams {
